Feat: scroll to new added field
